media-source/SourceBuffer-remove_invalid-updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Calling SourceBuffer#remove(start, end) during updating is true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var video = document.createElement('video');
var contents = {'/media/white.webm' : 'video/webm; codecs="vorbis,vp8"',
                '/media/segmented-3s.mp4'  : 'video/mp4'};

function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function testMedia(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updatestart', function(e) {
                    t.step(function() {
                        assert_throws({name: 'InvalidStateError'},
                                      function() {
                                          sourceBuffer.remove(0, 1);
                                      },
                                      'During updating is true, calling SourceBuffer#remove(start, end)');
                    });
                    t.done();
                });
                sourceBuffer.appendBuffer(data);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'SourceBuffer#remove(start, end) (' + mime + ') : ' +
       'During updating is true, SourceBuffer#remove(start, end) should throw InvalidStateError');
}
for(var file in contents) {
    testMedia(file, contents[file]);
}
</script>
</body>
</html>
